move handlers as a top level package
