+++ README.md
@@ -10,7 +10,7 @@ Collection of Logging helpers for use by HTTP services
 $ go get github.com/graze/golang-service/logging
 ```
 
-#### Healthd Logger
+### Healthd Logger
 
 - Support the healthd logs from AWS Elastic Beanstalk logs: [AWS](http://docs.aws.amazon.com/elasticbeanstalk/latest/dg/health-enhanced-serverlogs.html)
 
@@ -25,7 +25,7 @@ loggedRouter := logging.StatsdHandler(c, r)
 http.ListenAndServe(":1123", loggedRouter)
 ```
 
-#### Statsd Logger
+### Statsd Logger
 
 - Output `response_time` and `count` statistics for each request to a statsd host
 
@@ -40,27 +40,33 @@ func statsdHandler(h http.Handler) http.Handler {
 }
 ```
 
-### Pre-build handlers
+## Pre-build handlers
 
 This is a collection of pre-made handlers that use environment variables to quickly add to your application
 
-#### Syslog
+```bash
+$ go get github.com/graze/golang-service/handlers
+```
+
+### Syslog
 
 Output to syslog using the following environment variables
 
 Environment Variables:
+```
     SYSLOG_NETWORK: The network type of the syslog server (tcp, udp) Leave blank for local syslog
     SYSLOG_HOST: The host of the syslog server. Leave blank for local syslog
     SYSLOG_PORT: The port of the syslog server
     SYSLOG_APPLICATION: The application to report the logs as
     SYSLOG_LEVEL: The level to limit messages to (default: LEVEL6)
-
+```
 Example:
+```
     SYSLOG_NETWORK: udp
     SYSLOG_HOST: app.syslog.local
     SYSLOG_PORT: 1234
     SYSLOG_APPLICATION: app-live
-
+```
 Usage:
 ```go
 r := mux.NewRouter()
@@ -71,7 +77,7 @@ loggedRouter := handlers.SyslogHandler(r)
 http.ListenAndServe(":1123", loggedRouter)
 ```
 
-#### Healthd
+### Healthd
 
 This will create the application.log files in `/var/log/nginx/healthd/` and rotate them for you
 
@@ -80,22 +86,24 @@ loggedRouter := handlers.HealthdHandler(r)
 http.ListenAndServe(":1123", loggedRouter)
 ```
 
-#### StatsD
+### StatsD
 
 This will output to StatsD using the following variables
 
 Environment Variables:
+```
     STATSD_HOST: The host of the statsd server
     STATSD_PORT: The port of the statsd server
     STATSD_NAMESPACE: The namespace to prefix to every metric name
     STATSD_TAGS: A comma separared list of tags to apply to every metric reported
-
+```
 Example:
+```
     STATSD_HOST: localhost
     STATSD_PORT: 8125
     STATSD_NAMESPACE: app.live.
     STATSD_TAGS: env:live,version:1.0.2
-
+```
 Usage:
 ```go
 r := mux.NewRouter()
@@ -106,7 +114,7 @@ loggedRouter := handlers.StatsdHandler(r)
 http.ListenAndServe(":1123", loggedRouter)
 ```
 
-#### Combining
+### Combining
 
 These handlers can be combined in any combination or you can use the `handlers.AllHandlers(r)` as a short cut for all
 of them
