+++ handlers/syslog_test.go
